requirements.txt
# Core dependencies
streamlit==1.31.1
asyncio==3.4.3
python-dotenv==1.0.1

# HTTP and API dependencies
httpx==0.26.0
aiohttp==3.9.1

# Data handling
pyyaml==6.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1